add svg support
